+++ src/components/layout/SidebarPanel.tsx
@@ -10,6 +10,7 @@ import { type Component, Match, Switch } from 'solid-js'
 import { useLayout } from '../../stores/layout'
 import { ProjectSelector } from '../projects/ProjectSelector'
 import { SidebarExplorer } from '../sidebar/SidebarExplorer'
+import { SidebarPlugins } from '../sidebar/SidebarPlugins'
 import { SidebarSessions } from '../sidebar/SidebarSessions'
 
 export const SidebarPanel: Component = () => {
@@ -26,6 +27,9 @@ export const SidebarPanel: Component = () => {
         </Match>
         <Match when={activeActivity() === 'explorer'}>
           <SidebarExplorer />
+        </Match>
+        <Match when={activeActivity() === 'plugins'}>
+          <SidebarPlugins />
         </Match>
       </Switch>
     </aside>